+++ Extended desktop/Extended desktop.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29519.87
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Extended desktop", "Extended desktop\Extended desktop.vcxproj", "{D33EABA7-ECD3-44DC-9515-CEB942FEE71A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D33EABA7-ECD3-44DC-9515-CEB942FEE71A}.Debug|x64.ActiveCfg = Debug|x64
+		{D33EABA7-ECD3-44DC-9515-CEB942FEE71A}.Debug|x64.Build.0 = Debug|x64
+		{D33EABA7-ECD3-44DC-9515-CEB942FEE71A}.Debug|x86.ActiveCfg = Release|Win32
+		{D33EABA7-ECD3-44DC-9515-CEB942FEE71A}.Debug|x86.Build.0 = Release|Win32
+		{D33EABA7-ECD3-44DC-9515-CEB942FEE71A}.Release|x64.ActiveCfg = Release|x64
+		{D33EABA7-ECD3-44DC-9515-CEB942FEE71A}.Release|x64.Build.0 = Release|x64
+		{D33EABA7-ECD3-44DC-9515-CEB942FEE71A}.Release|x86.ActiveCfg = Release|Win32
+		{D33EABA7-ECD3-44DC-9515-CEB942FEE71A}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {95E82015-6EFC-44A9-8D44-A1BEC05550D7}
+	EndGlobalSection
+EndGlobal
