+++ Extended desktop/Extended desktop/Debug/Extended desktop.Build.CppClean.log
@@ -1,0 +1,12 @@
+c:\users\user\documents\github\extended_desktop\extended desktop\extended desktop\debug\vc142.pdb
+c:\users\user\documents\github\extended_desktop\extended desktop\extended desktop\debug\vc142.idb
+c:\users\user\documents\github\extended_desktop\extended desktop\extended desktop\debug\main.obj
+c:\users\user\documents\github\extended_desktop\extended desktop\debug\extended desktop.exe
+c:\users\user\documents\github\extended_desktop\extended desktop\debug\extended desktop.ilk
+c:\users\user\documents\github\extended_desktop\extended desktop\debug\extended desktop.pdb
+c:\users\user\documents\github\extended_desktop\extended desktop\extended desktop\debug\extended desktop.tlog\cl.command.1.tlog
+c:\users\user\documents\github\extended_desktop\extended desktop\extended desktop\debug\extended desktop.tlog\cl.read.1.tlog
+c:\users\user\documents\github\extended_desktop\extended desktop\extended desktop\debug\extended desktop.tlog\cl.write.1.tlog
+c:\users\user\documents\github\extended_desktop\extended desktop\extended desktop\debug\extended desktop.tlog\link.command.1.tlog
+c:\users\user\documents\github\extended_desktop\extended desktop\extended desktop\debug\extended desktop.tlog\link.read.1.tlog
+c:\users\user\documents\github\extended_desktop\extended desktop\extended desktop\debug\extended desktop.tlog\link.write.1.tlog
