+++ Extended desktop/Extended desktop/Debug/Extended desktop.log
@@ -1,0 +1,2 @@
+﻿  main.cpp
+  Extended desktop.vcxproj -> C:\Users\User\Documents\GitHub\extended_desktop\Extended desktop\Debug\Extended desktop.exe
