+++ Extended desktop/Extended desktop/Debug/Extended desktop.tlog/Extended desktop.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v142:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0
+Debug|Win32|C:\Users\User\Documents\GitHub\extended_desktop\Extended desktop\|
