+++ Extended desktop/Extended desktop/Release/Extended desktop.log
@@ -1,0 +1,7 @@
+﻿  main.cpp
+  Generating code
+  4 of 5 functions (80.0%) were compiled, the rest were copied from previous compilation.
+    0 functions were new in current compilation
+    0 functions had inline decision re-evaluated but remain unchanged
+  Finished generating code
+  Extended desktop.vcxproj -> C:\Users\User\Documents\GitHub\extended_desktop\Extended desktop\Release\Extended desktop.exe
