+++ Extended desktop/Extended desktop/Release/Extended desktop.tlog/Extended desktop.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v142:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0
+Release|Win32|C:\Users\User\Documents\GitHub\extended_desktop\Extended desktop\|
